Handle coupon discounts

+++ src/main/kotlin/rubiksworld/controller/database/tables/Coupons.kt
@@ -1,6 +1,7 @@
 package rubiksworld.controller.database.tables
 
 import org.ktorm.schema.Table
+import org.ktorm.schema.double
 import org.ktorm.schema.int
 import org.ktorm.schema.varchar
 import rubiksworld.model.Coupon
@@ -11,7 +12,7 @@ import rubiksworld.model.Coupon
 object Coupons : Table<Coupon>("Coupons") {
 
     val code = varchar("code").primaryKey().bindTo { it.code }
-    val value = int("value").bindTo { it.value }
+    val value = double("value").bindTo { it.value }
     val type = int("type")
         .transform({ Coupon.Type.values()[it] }, { it.ordinal })
         .bindTo { it.type }
